worker/requirements.txt
redis>=5.0.0
pymongo>=4.6.0
python-dotenv>=1.0.0
boto3>=1.34.0
yt-dlp>=2024.1.0
firebase-admin>=7.1.0
imageio-ffmpeg>=0.6.0
yt-dlp-get-pot
yt-dlp-get-pot-rustypipe
